backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
torch>=2.5.0
numpy>=1.26
gymnasium>=0.29.1
pandas>=2.2.0